install mysql only if it isnt installed

+++ index.js
@@ -23,6 +23,17 @@ function runSafe() {
   }
 }
 
+function checkInstalled(package) {
+  command = "dpkg -l | grep " + package;
+  const args = Array.from(arguments);
+  console.log(command);
+  let env = Object.assign({}, process.env);
+  delete env.CI; // for Homebrew on macos-11.0
+  const ret = execSync(command).toString();
+  console.log(ret);
+  return !ret || ret !== ""; 
+}
+
 function addToPath(newPath) {
   fs.appendFileSync(process.env.GITHUB_PATH, `${newPath}\n`);
 }
@@ -67,22 +78,25 @@ if (process.platform == 'darwin') {
     '5.7': '5.7.32',
     '5.6': '5.6.50'
   };
-  const fullVersion = versionMap[mysqlVersion];
-  useTmpDir();
-  run(`curl -Ls -o mysql.zip https://dev.mysql.com/get/Downloads/MySQL-${mysqlVersion}/mysql-${fullVersion}-winx64.zip`)
-  run(`unzip -q mysql.zip`);
-  fs.mkdirSync(`C:\\Program Files\\MySQL`);
-  fs.renameSync(`mysql-${fullVersion}-winx64`, `C:\\Program Files\\MySQL\\MySQL Server ${mysqlVersion}`);
+  installDir = "C:\\Program Files\\MySQL";
+  if (! fs.existsSync(installDir)) {
+    const fullVersion = versionMap[mysqlVersion];
+    useTmpDir();
+    run(`curl -Ls -o mysql.zip https://dev.mysql.com/get/Downloads/MySQL-${mysqlVersion}/mysql-${fullVersion}-winx64.zip`)
+    run(`unzip -q mysql.zip`);
+    fs.mkdirSync(installDir);
+    fs.renameSync(`mysql-${fullVersion}-winx64`, `C:\\Program Files\\MySQL\\MySQL Server ${mysqlVersion}`);
 
-  // start
-  bin = `C:\\Program Files\\MySQL\\MySQL Server ${mysqlVersion}\\bin`;
-  if (mysqlVersion != '5.6') {
-    run(`"${bin}\\mysqld" --initialize-insecure`);
+    // start
+    bin = `C:\\Program Files\\MySQL\\MySQL Server ${mysqlVersion}\\bin`;
+    if (mysqlVersion != '5.6') {
+      run(`"${bin}\\mysqld" --initialize-insecure`);
+    }
+    run(`"${bin}\\mysqld" --install`);
+    run(`net start MySQL`);
+
+    addToPath(bin);
   }
-  run(`"${bin}\\mysqld" --install`);
-  run(`net start MySQL`);
-
-  addToPath(bin);
 
   run(`"${bin}\\mysql" -u root -e "SELECT VERSION()"`);
 
@@ -91,25 +105,28 @@ if (process.platform == 'darwin') {
   run(`"${bin}\\mysql" -u root -e "GRANT ALL PRIVILEGES ON *.* TO 'ODBC'@'localhost'"`);
   run(`"${bin}\\mysql" -u root -e "FLUSH PRIVILEGES"`);
 } else {
-  if (image == 'ubuntu20') {
-    if (mysqlVersion != '8.0') {
-      // install
-      run(`sudo apt-get install mysql-server-${mysqlVersion}`);
-    }
-  } else {
-    if (mysqlVersion != '5.7') {
-      if (mysqlVersion == '5.6') {
-        throw `MySQL version not supported yet: ${mysqlVersion}`;
+  // check if it is installed
+  if(!checkInstalled("mysql-server-${mysqlVersion}")) {
+    if (image == 'ubuntu20') {
+      if (mysqlVersion != '8.0') {
+        // install
+        run(`sudo apt-get install mysql-server-${mysqlVersion}`);
       }
+    } else {
+      if (mysqlVersion != '5.7') {
+        if (mysqlVersion == '5.6') {
+          throw `MySQL version not supported yet: ${mysqlVersion}`;
+        }
 
-      // install
-      useTmpDir();
-      run(`wget -q -O mysql-apt-config.deb https://dev.mysql.com/get/mysql-apt-config_0.8.16-1_all.deb`);
-      run(`echo mysql-apt-config mysql-apt-config/select-server select mysql-${mysqlVersion} | sudo debconf-set-selections`);
-      run(`sudo dpkg -i mysql-apt-config.deb`);
-      // TODO only update single list
-      run(`sudo apt-get update`);
-      run(`sudo apt-get install mysql-server`);
+        // install
+        useTmpDir();
+        run(`wget -q -O mysql-apt-config.deb https://dev.mysql.com/get/mysql-apt-config_0.8.16-1_all.deb`);
+        run(`echo mysql-apt-config mysql-apt-config/select-server select mysql-${mysqlVersion} | sudo debconf-set-selections`);
+        run(`sudo dpkg -i mysql-apt-config.deb`);
+        // TODO only update single list
+        run(`sudo apt-get update`);
+        run(`sudo apt-get install mysql-server`);
+      }
     }
   }
